Unit support for number-text column (#1620)
## 🤨 Rationale

See #1573 

## 👩‍💻 Implementation

Clients configure a unit for a number-text column by providing a unit
element (e.g. `nimble-unit-byte`) as content. The unit element indicates
which unit-specific `UnitScale` to use. The `UnitScale` is used by the
column's formatter to pick an appropriate scaled unit for a given
number, and it provides a `UnitFormatter` to do the ultimate formatting.

There are two categories of `UnitScale`s: those that rely on
`Intl.NumberFormat` to provide translated unit labels, and those that
rely on translations built directly into Nimble.

Initial unit elements are `nimble-unit-byte` (which can be configured to
use either 1000-based or 1024-based units) and `nimble-unit-volt`. The
intent was to implement at least one of each category of `UnitScale` to
serve as examples for future unit implementations.

## 🧪 Testing

Unit tests created.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: mollykreis <20542556+mollykreis@users.noreply.github.com>
Co-authored-by: Jonathan Meyer <26874831+atmgrifter00@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>