Export Angular number formatting pipe (#1764)
# Pull Request

## 🤨 Rationale

Provides a way for clients to use the same formatting provided by the
number text table column in other parts of their apps.

## 👩‍💻 Implementation

- Created new pipe: `FormatNumberTextPipe` (using a corresponding new
`NumberTextUnitFormat` in `nimble-components`) with same configuration
options as the number-text table column
- Added documentation for new pipe in Storybook (number text column
story)
- Made the new pipe a standalone component and changed the existing
`DurationPipe` to be standalone as well. Removed the now empty
`NimblePipesModule`. This is a breaking change.

## 🧪 Testing

Tests added

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: rajsite <rajsite@users.noreply.github.com>