Performance tests for wafermap (#1786)
# Pull Request

## 🤨 Rationale

We wish to create performance tests using lighthouse, for that we are
going to create most of the infrustructure

## 👩‍💻 Implementation

The wafer performance tests will appear in a lighthouse report generated
by the CI after every push, or can be trigger manually using npm run
performance -w packages/performance.

At runtime, while we create/populate the wafer component we are
measuring the performance using windows.performance. The result will be
available at 'User Timing marks and measures'

To avoid confusion the existing lighthouse tests were moved in the
directory angular-workspace

The performance tests are built using vite

Despite lighthouse the landing site of the performance tests can be find
in the Chromatic build

The values on the wafer dies are computed simply (i + j) % 100, where i
and j are the coordinates of the die.

Also some math was used to compute the radius/sideLegth of the wafer.  
sideLength = Math.sqrt(diesCount / Math.PI) * 2;

it is derived from the circle's aria formula
A = π r²
where A in our case is the number of dies.

## 🧪 Testing

The results can be found in the CI

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: rajsite <rajsite@users.noreply.github.com>