Add basic support for delayed data hierarchy (#1794)
# Pull Request

## 🤨 Rationale

This is part of #1758

This PR adds support for claiming a row has children that are not
present in the data.

The ability to mark a row as loading its children will be included in a
follow up PR.

## 👩‍💻 Implementation

The majority of the implementation is adding the following based on [the
data hierarchy
HLD](https://github.com/ni/nimble/blob/main/packages/nimble-components/src/table/specs/data-hierarchy-hld.md#lazy-loading):
- `TableRecordHierarchyOptions` interface
- `TableRecordDelayedHierarchyState` enum (no `loadingChildren` state in
this PR)
- `setRecordHierarchyOptions` function on the table

I created a `RowOptionsManager` to track the options for the table. This
new manager, along with the `ExpansionManager` (which has a few
updates), have the majority of the logic for this feature.

## 🧪 Testing

- Manually tested in storybook
- Updated the matrix tests for the table to make one row have a state of
`canLoadChildren`
- Note: This update to the matrix tests exposed an existing bug with the
alignment of parent rows in Safari. In the version of Safari used in
Chromatic, the cell contents are 1px pushed to the right, but I wasn't
able to reproduce this locally in webkit. We can look into this as a
separate task if we want.
- Wrote new unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jonathan Meyer <26874831+atmgrifter00@users.noreply.github.com>