Workaround lighthouse build (#1809)
# Pull Request

## 🤨 Rationale

Disables the Card in the Angular and Blazor example apps as it breaks
axe-core tooling used by lighthouse, see:
https://github.com/ni/nimble/issues/1650

## 👩‍💻 Implementation

Disabled the nimble-card in the Angular and Blazor apps

Re-enabled the accessibility lighthouse error. The thought is that with
this change the accessibility score will be reliably calculated and not
cause intermittent blocking of merges, only blocks on merges that
further regress the score. If it is not the case and lighthouse becomes
unreliable calculating the score we can disable it again.

Removed the exisiting comment for an issue to get us back to 0.9 score
as there are multiple issues impacting our score now.

## 🧪 Testing

Relying on lighthouse CI execution.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.