Add publishConfig key to package files (#1826)
## 🤨 Rationale

For consistency, setting each package's npm `publishConfig` as public
access. This is necessary for the initial creation of the package, so if
we use any of these as a starting point for future packages, it's useful
to include. It also makes the intent explicit.

## 👩‍💻 Implementation

Added `publishConfig` setting to every package in Nimble repo, except
for `jasmine-parameterized`, which already had it.

## 🧪 Testing
N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>