Allow a delayed hierarchy row to be expanded before it has children (#1832)
# Pull Request

## 🤨 Rationale

Currently, a row is always considered collapsed when it doesn't have
children. However, this leads to some oddities when expanding a row with
delayed hierarchy and no children.

Therefore, a row should be expanded after the expand button is pressed
regardless of whether or not it has children present in the table yet.
But, a row should automatically collapse in the following circumstances:
1. It moves from the `loadingChildren` state to the `canLoadChildren`
state without having children.
2. It has a state of `canLoadChildren` and transitions from having child
rows in the data to having no child rows in the data.

See [this
discussion](https://github.com/ni/nimble/pull/1794#discussion_r1481996663)
for more context.

## 👩‍💻 Implementation

In order to handle the two circumstances mentioned above, I had to
refactor the table's expansion manager. The purpose of the refactor was
to:
- Allow each row to be in an explicitly expanded, explicitly collapsed,
or default state
- Purpose: The previous code in the expansion manager only tracked two
things: (1) whether or not every row was in a default state, and (2)
which rows were collapsed. This made it difficult to manage the state
when some rows defaulted to expanded and other to collapsed.
- Track which rows have children.
- Purpose: We want to collapse rows that transition into having no
children and rows without children that transition from the
`loadingChildren` state to the `canLoadChildren` state. Therefore, on a
data update, we need to know whether or not a row previously had
children, and when hierarchy options are set, we need to know the
current set of rows without children.

## 🧪 Testing

- Manually tested in storybook
- Wrote new unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>