Fix issue with underline-hidden on the anchor column (#1837)
# Pull Request

## 🤨 Rationale

This fixes an issue I noticed with the anchor column where the
`underline-hidden` configuration on the anchor column didn't always work
as expected. This is because the `underline-hidden` styling is based on
the presence of the `underline-hidden` attribute, but the cell-view
didn't use `?` when configuring `underline-hidden` on the underlying
`nimble-anchor` element.

## 👩‍💻 Implementation

Add `?` in the anchor column's cell-view template when setting
`underline-hidden` on the underlying `nimble-anchor`.

## 🧪 Testing

- Manually tested in storybook
- Wrote new unit test that fails without this change and passes with it

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.