Angular support for delayed table hierarchy (#1829)
# Pull Request

## 🤨 Rationale

This is part of #1758 

This PR adds support for delayed hierarchy within the table in Angular,
and it extends the Angular example app to have a table that has delayed
hierarchy.

## 👩‍💻 Implementation

- Expose necessary types from Angular table directive
- Add `row-loading` support to Angular's table label provider directives
- Extend the Angular example app to have a second table whose primary
purpose is to demonstrate the use of nimble's delayed hierarchy APIs

## 🧪 Testing

- Ran auto tests
- Manually used the new table in the example app

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.