Fix bug with indentation of nested parent rows (#1839)
# Pull Request

## 🤨 Rationale

This fixes a bug I found where the indentation of nested parent rows can
sometimes be incorrect, particularly when the order of rows changes
(such as when sorting a column) in a way that changes the nesting level
that is set on a `nimble-table-row` element that is getting reused
during a binding update.

The problematic binding was this condition in the table row's template:
`x.isParentRow && x.nestingLevel > 0`. That statement is volatile, but
it cannot be marked as such in the template. Therefore, I moved the
evaluation into a getter named `isNestedParent` that could be marked as
`@volatile`.

## 👩‍💻 Implementation

See _Rationale_ section above

## 🧪 Testing

Manually tested in storybook with a problematic set of data that I found
the problem with

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.