Card accessibility tweak (#1858)
# Pull Request

## 🤨 Rationale

The card was previously disabled after it started resulting in axe-core
completely failing. After reporting to axe-core their comment in
https://github.com/dequelabs/axe-core/issues/4335#issuecomment-1959646970
made me think we can workaround the axe-core issue by wrapping the slot
in a span and targeting that instead for aria.

While not necessary for runtime behavior or accessibility, the
workaround adds minimal overhead and complexity. Does not need to be
removed after the axe-core fix. Fixes #1650

## 👩‍💻 Implementation

Wraps the slot in a span and use that as the aria target. 

## 🧪 Testing

Can see the PR has a passing lighthouse score with the card re-enabled.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.