Fix bug with setting a title on spinner in Firefox (#1849)
# Pull Request

## 🤨 Rationale

Resolves #1825 

## 👩‍💻 Implementation

I created a transparent overlay on top of the spinning bits in the
spinner's template. This is enough to get the title to display correctly
in Firefox and not flicker when the bits move.

## 🧪 Testing

Manually verified that the title now displays correctly in Firefox and
still displays correctly in Edge.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.