Update usage of Intl.NumberFormat options (#1791)
# Pull Request

## 🤨 Rationale

Fixes: #1442 

## 👩‍💻 Implementation

The options cited in the issue are:

- `useGrouping: 'auto'` - We previously were setting this to `true`,
which means "display grouping separators even if the locale prefers
otherwise." Instead we want `"auto"`, which is "display grouping
separators based on the locale preference." The default for this option
is `"auto"` unless `notation` is `"compact"`. We never set `notation` to
`"compact"`, so we can just leave `useGrouping` to use the default.
- `signDisplay: 'negative'` - Though this option value has been
[supported in all browsers since
8/23](https://caniuse.com/mdn-javascript_builtins_intl_numberformat_numberformat_options_signdisplay_parameter_negative),
there does not seem to be a version of Typescript whose definition of
`Intl.NumberFormatOptions` allows it. There is an [open issue for adding
support](https://github.com/microsoft/TypeScript/issues/56269). As a
workaround, I've just added type assertions where necessary.
- `roundingPriority: 'lessPrecision'` - Just like `signDisplay:
'negative'`, the `roundingPriority` option has been [supported in
browsers since
8/23](https://caniuse.com/mdn-javascript_builtins_intl_numberformat_numberformat_options_roundingpriority_parameter),
but Typescript doesn't have support yet. It is tracked by the same issue
linked above, and our type assertion serves as a workaround for this,
too. By using this option, we can get rid of the separate
`leadingZeroFormatter` in `DefaultUnitFormat` and simplify the logic.

## 🧪 Testing

Existing tests pass.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.