Empty icon mapping (#1874)
# Pull Request

## 🤨 Rationale

Grouping on an icon value isn't ideal if any of the cells don't have an
icon. Specifically, the group row doesn't have a value associated with
it and renders only a row count.

It would be nice if we could allow a client to leave a cell blank for
certain field values but still have the group rows have a value
associated with them. For example, have a column that shows either a
spinner or nothing based on whether a task is in progress. When
grouping, it would be nice to show a string, such as "Idle", associated
with the records that have nothing in progress without cluttering the
table with extra icons in cells.

This is the Future Work described in #1869

## 👩‍💻 Implementation

Updated the icon column to allow a `nimble-mapping-icon` to have an
`undefined` icon. In that case, the cell renders nothing but the group
row renders only the text associated with the mapping.

## 🧪 Testing

Manually tested in storybook.
Added new unit tests for having an `undefined` icon
Updated storybook/matrix tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>