Fix drawer/select animation bug (#1871)
# Pull Request

## 🤨 Rationale

Fixes #1097

## 👩‍💻 Implementation

It turns out that when the drawer is opening from the right, and the
focused control in the drawer is a select, then the animation issue
occurs. The root cause is that as part of the FAST Listbox's
`focusAndScrollOptionIntoView` implementation, it queues
`scrollIntoView` on the selected option in a `requestAnimationFrame`
call. This interferes with the drawer's animation due to browser bugs
([Chromium](https://issues.chromium.org/issues/327554079),
[Firefox](https://bugzilla.mozilla.org/show_bug.cgi?id=1882886)).

We already override `focusAndScrollOptionIntoView` in our select, so
I've just moved the call to `super.focusAndScrollOptionIntoView` into an
`if (this.open)` block. This avoids (uneccessarily) calling
`scrollIntoView` while the animation is in progress.

## 🧪 Testing

I have added a test to ensure that option scrolling still happens
properly.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.