Fix anchor drag (#1910)
# Pull Request

## 🤨 Rationale

Fixes #1406 

## 👩‍💻 Implementation

The issue with the anchor items is that the slotted content was blocking
the pointer events (i.e. the drag attempt) and keeping the link from
being dragged. By styling the children of the native anchor elements
with `pointer-events: none;`, this is prevented and the link can be
dragged from all parts of the component.

## 🧪 Testing

Manually tested each changed component that the link can now be dragged
by the slotted content, including any start/end slot icons.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.