Add placeholder text to group rows (#1914)
# Pull Request

## 🤨 Rationale

This is part of #1538

This PR adds support for placeholder text in group rows within the
table. There is follow on work that will need to be done before the
linked issue is resolved to add placeholders to cells, but that is not
included in this PR.

This PR is scoped to the following behavior changes to columns:
- text column
    - Show "Empty" on group rows for empty string values
    - Show "No value" on group rows for `null` and `undefined` values
- date-text column
    - Show "No value" on group rows for `null` and `undefined` values
- number-text column
    - Show "No value" on group rows for `null` and `undefined` values
- duration-text column
    - Show "No value" on group rows for `null` and `undefined` values
- icon column
    - Show "No value" on group rows for `null` and `undefined` values
- enum-text column
    - Show "No value" on group rows for `null` and `undefined` values
- anchor column
    - Show "Empty" on group rows for empty label values
- Show "No value" on group rows for `null` and `undefined` label values

## 👩‍💻 Implementation

- Added new labels to the table label provider for _Empty_ and _No
value_
- Added a `applyPlaceholderTextIfNeeded` function to
`TableColumnTextGroupHeaderViewBase` that concrete classes can call to
apply a placeholder
- Update concrete instances of `TableColumnTextGroupHeaderViewBase` to
call the new function

## 🧪 Testing

- Wrote a number of new tests for each column to make sure the
appropriate group row text was rendered for different values
- Updated matrix tests to ensure each column included all possible
placeholder group values in the matrix test
- Note: This meant creating a new matrix story for the text column
because those matrix stories had never been created

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.