Fix intermittent RichTextMentionListbox test (#1928)
# Pull Request

## 🤨 Rationale

Fixes #1891 

## 👩‍💻 Implementation

Calling `show()` on the `RichTextMentionListbox` results in its anchored
region becoming un-hidden, and an intersection observer begins observing
the new anchor element (i.e. what the anchored region is attached to).
The observer closes the listbox when the anchor element is scrolled out
of view. Depending on the timing, the observer callback can run right
after calling `show()`, and if the anchor element is not fully visible
in the viewport, then the listbox can be closed before we get a chance
to run the rest of the test. Previously, the test had been using the
document element as the anchor element, but for some reason that I
couldn't discern, the intersection observer would say it was not
intersecting (with itself!). I changed the test to use a div (which
stays fully in the viewport) as the anchor element, and we no longer run
into an issue with the intersection observer prematurely closing the
listbox.

## 🧪 Testing

Interestingly, focusing the intermittent test with `fit` would cause it
to always succeed. I had to disable all other tests (with `xdescribe`)
to run the flaky test by itself. I ran the test with my changes many
times without any failures.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>