Create storybook page for link components (#1945)
# Pull Request

## 🤨 Rationale

There are a number of browser interactions that are expected to work in
certain ways for link components that are difficult to test in an
automated manner. Therefore, we decided that we should have a storybook
page that contains all of nimble's link components next to a native
anchor element to facilitate manual testing of link behaviors.
 
This is follow-on work from #1910, specifically [this
comment](https://github.com/ni/nimble/pull/1910#discussion_r1518327840).

## 👩‍💻 Implementation

Create storybook page that includes the following elements:
- native anchor
- nimble-anchor
- nimble-anchor-button
- nimble-anchor-tabs
- nimble-breadcrumb
- nimble-anchor-tree-item
- nimble-anchor-menu-item
- nimble-table-column-anchor
- nimble-rich-text-viewer

## 🧪 Testing

Manually verified that the storybook page rendered as expected.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>