Manually update Plawright to 1.42.0 (#1956)
# Pull Request

## 🤨 Rationale

#1747 has a recurring maintenance task to update the version of
Playwright to the latest version that's available in npm and Nuget. We
were on 1.40.0, the [latest on npm is
1.42.1](https://www.npmjs.com/package/playwright?activeTab=versions),
and the [latest on Nuget is
1.42.0](https://www.nuget.org/packages/Microsoft.Playwright)

[Detailed release
notes](https://github.com/microsoft/playwright/releases).

## 👩‍💻 Implementation

1. Searched for "1.40.0" in package.json files and updated to "1.42.0".
2. Ran `npm install` to regenerate package-lock.json

## 🧪 Testing

1. Verified locally that
`packages/nimble-blazor/build/generate-playwright-version-properties/dist/Playwright.PackageVersion.props`
is regenerated and the build and test commands pass.
2. Otherwise relying on PR build

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.