Created Computations and Data Manager Experimental (#1946)
# Pull Request

## 🤨 Rationale

Created new computations experimental component with changed scales and
diesTable iteration for finding the bounding box until arquero fixes.

The new scales will be able to be inverted without the need to create
separate scales for the hover handler.

## 👩‍💻 Implementation

changed the scales and iteration over the diesTable
integrated the changes and the strategy swap with the wafermap and the
experimental hover

## 🧪 Testing

copied spec classes for the new experimental modules.
updated hover handler spec.
adapted the data manager to the previous renderer for visual tests in
storybook.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.