Add cell placeholders (#1950)
# Pull Request

## 🤨 Rationale

This PR adds a `placeholder` configuration to the following table
columns as described in #1869:
- anchor
- date-text
- duration-text
- number-text
- text

This is part of #1538. The remaining work is to add the `placeholder`
attributes to the Angular and Blazor wrappers for each column.

As part of this PR, I also included best practices established for each
column in the storybook docs. These are adapted from the placeholder
HLD.

## 👩‍💻 Implementation

I implemented this change in a similar way as #1914. Specifically, I
updated `TableColumnTextCellViewBase` to have its own implementation of
`columnConfigChanged` and `cellRecordChanged` that calls a new
`applyPlaceholderTextIfNeeded` function. This required me to create a
`TableColumnTextBaseCellRecord` interface to enforce that the cell
records used by `TableColumnTextCellViewBase` have a `value` property
and to create a `TableColumnTextBaseColumnConfig` interface to enforce
that the column configs used by `TableColumnTextCellViewBase` have an
optional `placeholder` property.

## 🧪 Testing

- Updated/created unit tests
- Manually tested in storybook
- Updated matrix tests to test each relevant column with & without a
placeholder

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>