Update chromatic config to be from root (#1965)
# Pull Request

## 🤨 Rationale

Chromatic was failing to detect the storybookdir configuration
correctly. Guess was that the interactions between workingdir and the
other options were weird. Removed the working dir option and kept all
paths relative to the repo root.

## 👩‍💻 Implementation

See above.

## 🧪 Testing

Validated in a build that turbosnap detected just changes associated
with updated files:

https://github.com/ni/nimble/actions/runs/8456254529/job/23165615846?pr=1965#step:13:52

Validated that disabling turbosnap for the branch (as if running on
main) seemed to upload correctly:

https://github.com/ni/nimble/actions/runs/8456510572/job/23166454906#step:13:61

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.