Fix table scroll tests on WebKit (#1963)
# Pull Request

## 🤨 Rationale

Resolves #1942 

## 👩‍💻 Implementation

WebKit doesn't appear to fire the `scroll` event when a scroll done by
setting `scrollTop` on an element. Therefore, I've updated the page
object to scroll using the `scroll()` function, which does cause the
event to be fired. As a result, the table's virtualizer correctly
detects the change and the tests pass.

## 🧪 Testing

Ran all the table tests in webkit locally and on the pipeline to verify
they pass.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>