Set build time zone to America/Chicago (#1968)
# Pull Request

## 🤨 Rationale

This came up in [a discussion on the PR to add placeholders to table
columns](https://github.com/ni/nimble/pull/1950#discussion_r1538266790).
The goal is the increase pipeline stability and reproducibility by
ensuring more of the configuration is consistent regardless of what
machine is running the pipeline. In this case, I am setting the time
zone to _America/Chicago_.

## 👩‍💻 Implementation

- set `TZ` environment variable to `Americal/Chicago`
- set `--time-zone-for-testing=America/Chicago` in the Chrome flags for
the nimble-components unit tests

## 🧪 Testing

Build succeeds
Verified locally that the `--time-zone-for-testing` Chrome flag is used
as expected

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.